requirements.txt
streamlit==1.33.0
streamlit-webrtc==0.47.6
openai==1.6.1
python-dotenv==1.0.0
opencv-python==4.9.0.80
numpy==1.26.4
scipy==1.11.4
python-socketio==5.10.0
Pillow==10.1.0
pydub==0.25.1
SpeechRecognition==3.10.0
pyaudio==0.2.14
gtts==2.5.0
requests==2.31.0
openai-whisper==20231117
torch>=2.0.0
torchaudio>=2.0.0
reportlab==4.0.7
fpdf==1.7.2
